bring the VC7 build up to date and disable a lot of warnings
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@395 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ src/device_ds_stream.cpp
@@ -150,7 +150,7 @@ namespace audiere {
 
   void
   DSOutputStream::setPitchShift(float shift) {
-    m_buffer->SetFrequency(m_base_frequency * shift);
+    m_buffer->SetFrequency(DWORD(m_base_frequency * shift));
   }
